.
- bump descriptor pool size
- limit deferred lights to 16 atm
- frameytime

+++ lvk/include/VulkanAPI.h
@@ -319,7 +319,7 @@ public:
         // API End
     protected:
         bool            p_ShouldRun = true;
-        double          p_LastFrameTime;
+        uint64_t        p_LastFrameTime;
         int             p_CurrentFrameIndex;
     };
 }
